+++ client/src/config.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 use serde::Deserialize;
 use serde_with::{
   base64::{Base64, Standard},
