+++ core/fixtures/claude_shannon.html
