+++ core/src/http.rs
@@ -214,7 +214,7 @@ impl NotaryResponse {
   fn body_matches(
     &self,
     other: &ManifestResponse,
-  ) -> Result<Option<ExtractionValues>, ManifestError> {
+  ) -> Result<Option<ExtractionValues>, WebProverCoreError> {
     match &self.notary_response_body.body {
       Some(body) => {
         let result = other.body.0.extract_and_validate(body)?;
@@ -921,249 +921,5 @@ pub mod tests {
     });
 
     assert!(request.validate_vars().is_ok());
-  }
-
-  // Creates a response body with a given serde_json::Value
-  #[macro_export]
-  macro_rules! notary_body {
-    ($($key:tt : $value:tt),* $(,)?) => {{
-        NotaryResponseBody {
-            json: Some(json!({
-                $(
-                    $key: $value
-                ),*
-            }))
-        }
-    }};
-}
-
-  #[test]
-  fn test_json_contains_path_valid_object_path() {
-    let json = notary_body!(
-        "key1": {
-            "key2": {
-                "key3": "value"
-            }
-        }
-    );
-    // ["key1", "key2", "key3"]
-    let path = vec![
-      JsonKey::String("key1".to_string()),
-      JsonKey::String("key2".to_string()),
-      JsonKey::String("key3".to_string()),
-    ];
-    assert!(json.matches_path(&path));
-
-    // ["key1", "key2", "key3", "value"]
-    let mut path_with_value = path.clone();
-    path_with_value.push(JsonKey::String("value".to_string()));
-    assert!(json.matches_path(&path_with_value));
-
-    // ["key1", "key2", "key3", "value", "invalid"]
-    let mut path_with_invalid_value = path.clone();
-    path_with_invalid_value.push(JsonKey::String("invalid_value".to_string()));
-    assert!(!json.matches_path(&path_with_invalid_value));
-  }
-
-  #[test]
-  fn test_json_contains_path_valid_array_path() {
-    let json = notary_body!(
-        "key1": [
-            {
-                "key2": "value1"
-            },
-            {
-                "key3": "value2"
-            }
-        ]
-    );
-    let path = vec![
-      JsonKey::String("key1".to_string()),
-      JsonKey::Num(1),
-      JsonKey::String("key3".to_string()),
-      JsonKey::String("value2".to_string()),
-    ];
-    assert!(json.matches_path(&path));
-  }
-
-  #[test]
-  fn test_json_contains_path_invalid_key() {
-    let json = notary_body!(
-        "key1": {
-            "key2": "value"
-        }
-    );
-    let path =
-      vec![JsonKey::String("key1".to_string()), JsonKey::String("invalid_key".to_string())];
-    assert!(!json.matches_path(&path));
-  }
-
-  #[test]
-  fn test_json_contains_path_invalid_array_index() {
-    let json = notary_body!(
-        "key1": [
-            "value1",
-            "value2"
-        ]
-    );
-    let path = vec![
-      JsonKey::String("key1".to_string()),
-      JsonKey::Num(3), // Out of bounds
-    ];
-    assert!(!json.matches_path(&path));
-  }
-
-  #[test]
-  fn test_json_contains_path_index_on_non_array() {
-    let json = notary_body!(
-        "key1": {
-            "key2": "value"
-        }
-    );
-    let path = vec![
-      JsonKey::String("key1".to_string()),
-      JsonKey::Num(0), // Applying index on a non-array
-    ];
-    assert!(!&json.matches_path(&path));
-  }
-
-  #[test]
-  fn test_empty_path() {
-    let json = notary_body!(
-        "key1": {
-            "key2": {
-                "key3": "value"
-            }
-        }
-    );
-    let path: Vec<JsonKey> = vec![];
-    assert!(!json.matches_path(&path));
-  }
-
-  #[test]
-  fn test_empty_body() {
-    let json = notary_body!();
-    let mut path: Vec<JsonKey> = vec![];
-    assert!(!json.matches_path(&path));
-
-    path.push(JsonKey::String("key1".to_string()));
-    assert!(!json.matches_path(&path));
-  }
-
-  #[test]
-  fn test_number_matching() {
-    let test_cases = vec![
-      // Unsigned integers
-      (json!({"val": 42u64}), 42, true),
-      // Signed integers
-      (json!({"val": -42}), -42, true),
-      // Floating point
-      (json!({"val": 42.0}), 42, true),
-      // Large numbers
-      (json!({"val": u64::MAX}), u64::MAX as i64, true),
-      // Mismatched types
-      (json!({"val": "42"}), 42, false),
-      // Edge cases
-      (json!({"val": 42.000001}), 42, false),
-    ];
-
-    for (json, expected, should_match) in test_cases {
-      let body = NotaryResponseBody { json: Some(json.clone()) };
-      assert_eq!(
-        body.matches_path(&[JsonKey::String("val".to_string()), JsonKey::Num(expected as usize)]),
-        should_match,
-        "Failed for value: {:?}, expected: {}",
-        json,
-        expected
-      );
-    }
-  }
-
-  #[test]
-  fn test_nested_array_path() {
-    let json = notary_body!(
-        "data": [
-            [1, 2, 3],
-            [4, 5, 6]
-        ]
-    );
-    assert!(json.matches_path(&[
-      JsonKey::String("data".to_string()),
-      JsonKey::Num(1),
-      JsonKey::Num(1),
-      JsonKey::Num(5),
-    ]));
-  }
-
-  #[test]
-  fn test_mixed_types_in_array() {
-    let json = notary_body!(
-        "mixed": [
-            42,
-            "string",
-            { "key": "value" },
-            [1, 2, 3]
-        ]
-    );
-    assert!(json.matches_path(&[
-      JsonKey::String("mixed".to_string()),
-      JsonKey::Num(0),
-      JsonKey::Num(42),
-    ]));
-    assert!(json.matches_path(&[
-      JsonKey::String("mixed".to_string()),
-      JsonKey::Num(1),
-      JsonKey::String("string".to_string()),
-    ]));
-    assert!(json.matches_path(&[
-      JsonKey::String("mixed".to_string()),
-      JsonKey::Num(2),
-      JsonKey::String("key".to_string()),
-      JsonKey::String("value".to_string()),
-    ]));
-  }
-
-  #[test]
-  fn test_empty_string_keys_and_values() {
-    let json = notary_body!(
-        "": {
-            "empty": ""
-        }
-    );
-    // Test empty string keys and values
-    assert!(json.matches_path(&[
-      JsonKey::String("".to_string()),
-      JsonKey::String("empty".to_string()),
-      JsonKey::String("".to_string()),
-    ]));
-  }
-
-  #[test]
-  fn test_very_deep_nesting() {
-    // Test different depths: 5, 25, and 100 levels
-    let depths = vec![5, 25, 100];
-
-    for depth in depths {
-      // Build nested JSON structure
-      let mut json_value = json!("value");
-      for i in (0..depth).rev() {
-        json_value = json!({
-            format!("level{}", i): json_value
-        });
-      }
-
-      let json = NotaryResponseBody { json: Some(json_value) };
-
-      // Build the path
-      let mut path = Vec::new();
-      for i in 0..depth {
-        path.push(JsonKey::String(format!("level{}", i)));
-      }
-      path.push(JsonKey::String("value".to_string()));
-
-      // Test the path
-      let result = json.matches_path(&path);
-      assert!(result, "Failed to match path at depth {}: {:?}", depth, path);
-    }
   }
 }
