+++ core/src/parser/mod.rs
@@ -1,13 +1,14 @@
-mod extractor;
-
-mod common;
 mod config;
 mod errors;
 mod predicate;
 
+mod extractors;
 #[cfg(test)] mod test_fixtures;
 #[cfg(test)] mod test_utils;
 
 pub use config::{DataFormat, ExtractorConfig};
 pub use errors::ExtractorError;
-pub use extractor::{ExtractionValues, Extractor, ExtractorType};
+pub use extractors::{
+  DocumentExtractor, ExtractionResult, ExtractionValues, Extractor, ExtractorType,
+  HtmlDocumentExtractor, JsonDocumentExtractor,
+};
