+++ core/src/parser/test_utils.rs
@@ -1,7 +1,10 @@
+use std::fs;
+
 use serde_json::{json, Value};
+use tl::{ParserOptions, VDom};
 
 use crate::parser::{
-  extractor::ExtractionResult,
+  extractors::ExtractionResult,
   predicate::{Comparison, PredicateType},
   DataFormat, Extractor, ExtractorConfig, ExtractorType,
 };
@@ -22,6 +25,7 @@ macro_rules! extractor {
             extractor_type: crate::parser::ExtractorType::String,
             required: true,
             predicates: Vec::new(),
+            attribute: None,
         };
 
         // Override default fields with provided arguments
@@ -139,4 +143,12 @@ pub fn tags_extractor() -> Extractor {
       value: json!(1)
     )]
   )
+}
+
+pub fn create_complex_test_html() -> String {
+  fs::read_to_string("../core/fixtures/website.html").unwrap()
+}
+
+pub fn parse_html(html: &str) -> VDom {
+  tl::parse(html, ParserOptions::default()).expect("Failed to parse HTML")
 }
